First Commit
This project was intended to provide OO solution to a CO dev Test. The main goal is provided using a class for the CarInsurance  and a method with switch case to evaluate all the product rules. It was tested using mocha, coverage with istanbul. The proyect was ready to build, linted and optionaly has a log folder with winston. The log was deactivate and intended only for development in this project